refactor passes unit tests
