--- image_info/__init__.py
@@ -1,4 +1,0 @@
-"""
-The image_info module contains python that fetches the image information that
-populates the database
-"""
